refactor: Split the settings models and preview samples
Both files were over the size the linter allows. The models are now
grouped by what they describe, and the preview samples follow the
same grouping.
